fixed td3 tets and renamed util
