+++ qdax/baselines/sac_pbt.py
@@ -21,7 +21,7 @@ from qdax.core.neuroevolution.buffers.buffer import (
     Transition,
 )
 from qdax.core.neuroevolution.normalization_utils import normalize_with_rmstd
-from qdax.core.neuroevolution.sac_utils import do_iteration_fn
+from qdax.core.neuroevolution.sac_td3_utils import do_iteration_fn
 from qdax.types import Descriptor, Mask, Metrics, RNGKey
