+++ qdax/baselines/td3.py
@@ -22,7 +22,7 @@ from qdax.core.neuroevolution.losses.td3_loss import (
 )
 from qdax.core.neuroevolution.mdp_utils import TrainingState, get_first_episode
 from qdax.core.neuroevolution.networks.td3_networks import make_td3_networks
-from qdax.core.neuroevolution.sac_utils import generate_unroll
+from qdax.core.neuroevolution.sac_td3_utils import generate_unroll
 from qdax.types import (
     Action,
     Descriptor,
